refactor(auth): move generateDefaultUserFromSession to authService and remove userCreationHelper
